Change coordinate system to float to match collision coordinates

+++ core/src/Model/Enemies/Enemy.java
@@ -1,14 +1,13 @@
 package Model.Enemies;
 
 import Model.Entity;
-import Model.IModel;
 import com.dongbat.jbump.World;
 
 public abstract class Enemy extends Entity {
     private float damage;
     private String imagePath = ""; // hur gör man att de måste finnas en?
 
-    public Enemy(int x, int y, int height, int width, int speed, float health, float damage, World<Entity> world) {
+    public Enemy(float x, float y, float height, float width, float speed, float health, float damage, World<Entity> world) {
         super(x, y, height, width, speed, health, world);
         this.damage = damage;
     }
